Fix numerous VC-10 compiler warnings. Rewrite ICU configuration to use the new Boost.Build configuration logic.
git-svn-id: http://svn.boost.org/svn/boost/trunk@61893 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/regex/test/c_compiler_checks/posix_api_check.c
@@ -35,7 +35,7 @@ int main()
    regex_tA re;
    int result;
    result = regcompA(&re, expression, REG_AWK);
-   if(result > REG_NOERROR)
+   if(result > (int)REG_NOERROR)
    {
       char buf[256];
       regerrorA(result, &re, buf, sizeof(buf));
@@ -46,7 +46,7 @@ int main()
    matches[0].rm_so = 0;
    matches[0].rm_eo = strlen(text);
    result = regexecA(&re, text, 1, matches, REG_NOTBOL | REG_NOTEOL | REG_STARTEND);
-   if(result > REG_NOERROR)
+   if(result > (int)REG_NOERROR)
    {
       char buf[256];
       regerrorA(result, &re, buf, sizeof(buf));
